fix: make chatId `sql.BigInt` (#345)

+++ src/dal/Mssql/MssqlDataOperations.ts
@@ -23,14 +23,14 @@ export class MssqlDataOperations implements DataOperations {
     public async SetPlayerChatId(playerId: number, chatId: number): Promise<void> {
         const request = await this.getRequest();
         request.input('playerId', sql.Int, playerId);
-        request.input('chatId', sql.Int, chatId);
+        request.input('chatId', sql.BigInt, chatId);
         await request.execute<Player>('telegram.SetPlayerChatId');
         return;
     }
 
     public async GetPlayerFromChatId(chatId: number): Promise<Player> {
         const request = await this.getRequest();
-        request.input('chatId', sql.Int, chatId);
+        request.input('chatId', sql.BigInt, chatId);
         const result = await request.execute<Player>('telegram.GetPlayerFromChatId');
         return result.recordsets[0][0];
     }
